+++ src/valency_anndata/datasets/__init__.py
@@ -1,10 +1,12 @@
 from .polis import load, translate_statements
 from ._load_aufstehen import aufstehen
 from ._load_chile_protest import chile_protest
+from ._load_klimarat import klimarat
 
 __all__ = [
     "load",
     "aufstehen",
     "chile_protest",
+    "klimarat",
     "translate_statements",
 ]